+++ lib/mutant/isolation/fork.rb
@@ -3,12 +3,36 @@
 module Mutant
   class Isolation
     # Isolation via the fork(2) systemcall.
+    #
+    # Communication between parent and child process is done
+    # via anonymous pipes.
+    #
+    # Timeouts are initially handled relatively efficiently via IO.select
+    # but once the child process pipes are on eof via busy looping on
+    # waitpid2 with Process::WNOHANG set.
+    #
+    # Handling timeouts this way is not the conceptually most
+    # efficient solution. But its cross platform.
+    #
+    # Design constraints:
+    #
+    # * Support Linux
+    # * Support MacOSX
+    # * Avoid platform specific APIs and code.
+    # * Only use ruby corelib.
+    # * Do not use any named resource.
+    # * Never block on latency inducing systemcall without a
+    #   timeout.
+    # * Child process freezing before closing the pipes needs to
+    #   be detected by parent process.
+    # * Child process freezing after closing the pipes needs to be
+    #   detected by parent process.
     class Fork < self
       include(Adamantium::Flat, Concord.new(:world))
 
       READ_SIZE = 4096
 
-      ATTRIBUTES = %i[block log_pipe result_pipe world].freeze
+      ATTRIBUTES = %i[block deadline log_pipe result_pipe world].freeze
 
       # Unsuccessful result as child exited nonzero
       class ChildError < Result
@@ -64,9 +88,11 @@ module Mutant
         #
         # @return [Result]
         def call
-          pid = start_child or return ForkError.new
+          @log_fragments = []
 
-          read_child_result(pid)
+          @pid = start_child or return ForkError.new
+
+          read_child_result
 
           @result
         end
@@ -85,30 +111,45 @@ module Mutant
         end
 
         # rubocop:disable Metrics/MethodLength
-        def read_child_result(pid)
+        def read_child_result
           result_fragments = []
-          log_fragments    = []
 
-          read_fragments(
-            log_pipe.parent    => log_fragments,
-            result_pipe.parent => result_fragments
-          )
+          targets =
+            {
+              log_pipe.parent    => @log_fragments,
+              result_pipe.parent => result_fragments
+            }
 
-          begin
-            result = world.marshal.load(result_fragments.join)
-          rescue ArgumentError => exception
-            add_result(Result::Exception.new(exception))
+          read_targets(targets)
+
+          if targets.empty?
+            read_result(result_fragments)
+            terminate_graceful
           else
-            add_result(Result::Success.new(result, log_fragments.join))
+            add_result(Result::Timeout.new(deadline.allowed_time))
+            terminate_ungraceful
           end
-        ensure
-          wait_child(pid, log_fragments)
         end
         # rubocop:enable Metrics/MethodLength
 
-        def read_fragments(targets)
+        def read_result(result_fragments)
+          result = world.marshal.load(result_fragments.join)
+        rescue ArgumentError => exception
+          add_result(Result::Exception.new(exception))
+        else
+          add_result(Result::Success.new(result, @log_fragments.join))
+        end
+
+        # rubocop:disable Metrics/MethodLength
+        def read_targets(targets)
           until targets.empty?
-            ready, = world.io.select(targets.keys)
+            status = deadline.status
+
+            break unless status.ok?
+
+            ready, = world.io.select(targets.keys, [], [], status.time_left)
+
+            break unless ready
 
             ready.each do |fd|
               if fd.eof?
@@ -119,13 +160,43 @@ module Mutant
             end
           end
         end
+        # rubocop:enable Metrics/MethodLength
 
-        def wait_child(pid, log_fragments)
-          _pid, status = world.process.wait2(pid)
+        # rubocop:disable Metrics/MethodLength
+        def terminate_graceful
+          status = nil
 
-          unless status.success? # rubocop:disable Style/GuardClause
-            add_result(ChildError.new(status, log_fragments.join))
+          loop do
+            status = peek_child
+            break if status || deadline.expired?
+            world.kernel.sleep(0.1)
           end
+
+          if status
+            handle_status(status)
+          else
+            terminate_ungraceful
+          end
+        end
+        # rubocop:enable Metrics/MethodLength
+
+        def terminate_ungraceful
+          world.process.kill('KILL', @pid)
+
+          _pid, status = world.process.wait2(@pid)
+
+          handle_status(status)
+        end
+
+        def handle_status(status)
+          unless status.success? # rubocop:disable Style/GuardClause
+            add_result(ChildError.new(status, @log_fragments.join))
+          end
+        end
+
+        def peek_child
+          _pid, status = world.process.wait2(@pid, Process::WNOHANG)
+          status
         end
 
         def add_result(result)
@@ -162,12 +233,14 @@ module Mutant
       # ignore :reek:NestedIterators
       #
       # rubocop:disable Metrics/MethodLength
-      def call(&block)
+      def call(timeout, &block)
+        deadline = world.deadline(timeout)
         io = world.io
         Pipe.with(io) do |result|
           Pipe.with(io) do |log|
             Parent.call(
               block:       block,
+              deadline:    deadline,
               log_pipe:    log,
               result_pipe: result,
               world:       world
@@ -176,6 +249,7 @@ module Mutant
         end
       end
       # rubocop:enable Metrics/MethodLength
+
     end # Fork
   end # Isolation
 end # Mutant