+++ lib/mutant/world.rb
@@ -20,6 +20,7 @@ module Mutant
       :stderr,
       :stdout,
       :thread,
+      :timer,
       :warnings
     )
 
@@ -46,6 +47,20 @@ module Mutant
         Either::Right.new(stdout)
       else
         Either::Left.new("Command #{command} failed!")
+      end
+    end
+
+    # Deadline
+    #
+    # @param [Float, nil] allowed_time
+    def deadline(allowed_time)
+      if allowed_time
+        Timer::Deadline.new(
+          allowed_time: allowed_time,
+          timer:        timer
+        )
+      else
+        Timer::Deadline::None.new
       end
     end
   end # World